+++ src/vue/pages/dashboard/Dashboard.AssetSelector.vue
@@ -87,7 +87,7 @@ import NoDataMessage from '@/vue/common/NoDataMessage'
 import { ASSET_POLICIES } from '@tokend/js-sdk'
 import { mapGetters, mapActions } from 'vuex'
 import { vuexTypes } from '@/vuex'
-import { api } from '@/api'
+import { api, documentsManager } from '@/api'
 import { ErrorHandler } from '@/js/helpers/error-handler'
 import { AssetRecord } from '@/js/records/entities/asset.record'
 
@@ -151,7 +151,7 @@ export default {
     },
     imgUrl () {
       const balance = this.balances.find(i => i.asset === this.currentAsset)
-      return balance.assetDetails.logoUrl(config.FILE_STORAGE)
+      return documentsManager.getDocumentUrlByKey(balance.assetDetails.logoKey)
     },
   },
   async created () {